requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9
langchain>=1.1.0
langgraph>=1.0.4
databricks-langchain>=0.11.0
databricks-sdk>=0.73.0
databricks-vectorsearch>=0.40
mlflow>=3.6.0
pydantic>=2.12.5
pydantic-settings>=2.7.0
python-jose[cryptography]>=3.3.0
structlog>=24.0.0
psutil>=6.0.0
python-dotenv>=1.0.0
